% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ecampus_colors.R
\name{get_ecampus_colors}
\alias{get_ecampus_colors}
\title{Get ECampus Colors by Name}
\usage{
get_ecampus_colors(...)
}
\arguments{
\item{...}{One or more color names as character strings.}
}
\value{
A named vector of hex codes for the requested colors.
}
\description{
This function retrieves one or more color hex codes from the \code{ecampus_palettes} dataset.
}
\examples{
get_ecampus_colors(c("eCore_Green", "eCore_Teal"))
get_ecampus_colors("eCore_OER_Orange", palette = "USG-eCore")
get_ecampus_colors("eMajor_Navy", palette = "all") # Search in all branches
}
