% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ecampus_colors.R
\name{get_ecampus_colors}
\alias{get_ecampus_colors}
\title{Retrieve Colors from eCampus Palettes}
\usage{
get_ecampus_colors(...)
}
\arguments{
\item{...}{One or more color names (as strings) from the eCampus palettes.}
}
\value{
A named character vector containing the color values corresponding to the provided color names.
}
\description{
This function retrieves color values from the eCampus palettes based on the provided color names.
You can pass a single color name or multiple color names as arguments. The function will return
the corresponding color values from the \code{ecampus_palette}.
}
\examples{
# Example 1: Retrieve a single color
get_ecampus_colors("eCore_Teal")

# Example 2: Retrieve multiple colors
get_ecampus_colors("eMajor_Navy", "eCore_Teal", "eMajor_Pedagogy_Purple")

}
